feat(feature/stats): add Stats for SuperAgent
